requirements_boxmot.txt
# BoxMOT Integration Requirements
# Core tracking library
boxmot>=13.0.0

# Computer vision and image processing
opencv-python>=4.5.0
numpy>=1.21.0

# Deep learning framework
torch>=1.9.0
torchvision>=0.10.0

# Optional: For YOLO integration
ultralytics>=8.0.0

# Optional: For advanced ReID models
onnx>=1.12.0
onnxruntime>=1.12.0

# Utilities
pathlib2>=2.3.0
pyyaml>=6.0
tqdm>=4.64.0

# Optional: For GPU acceleration
# torch-audio  # Uncomment if needed
# torchaudio   # Uncomment if needed

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0 